requirements.txt
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
Django==2.0
future==0.16.0
idna==2.6
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-telegram-bot==9.0.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
